requirements.txt
asgiref==3.8.1
Django==5.2
django-rest-knox==5.0.2
django-utils==0.0.2
djangorestframework==3.16.0
pillow==11.1.0
sqlparse==0.5.3
